+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## v0.1.0
 - Initial version
+- Add `match_ordered_array` matcher